+++ dev_zaplane/redux/Slices/logsSlice/logsSlice.js
@@ -21,7 +21,7 @@ export const getSingleRunDetails = createAsyncThunk(
 				namespace + `runs/${parseInt(runId)}`
 			);
 
-			handleSliceSuccess(thunkAPI, __('Run details fetched successfully', 'workflow'));
+			handleSliceSuccess(thunkAPI, __('Run details fetched successfully', 'zaplane'));
 
 			return res.data;
 			// {
@@ -72,7 +72,7 @@ export const retryNodeRun = createAsyncThunk(
 
       thunkAPI.dispatch(
         showNotification({
-          message: __('Node retried and queued successfully', 'workflow'),
+          message: __('Node retried and queued successfully', 'zaplane'),
           isShow: true,
           type: 'success',
         })
@@ -97,7 +97,7 @@ export const clearRuns = createAsyncThunk(
 
       thunkAPI.dispatch(
         showNotification({
-          message: __('Logs cleared successfully', 'workflow'),
+          message: __('Logs cleared successfully', 'zaplane'),
           isShow: true,
           type: 'success',
         })
@@ -118,7 +118,7 @@ export const deleteRun = createAsyncThunk(
 
       thunkAPI.dispatch(
         showNotification({
-          message: __('Log deleted successfully', 'workflow'),
+          message: __('Log deleted successfully', 'zaplane'),
           isShow: true,
           type: 'success',
         })